refactor(zumabot): refactor and organize files
