+++ bots/zumabot/data/requirements.txt
